labels: fix comparison
The implementation of `less` properly sees that \x01 < \xfe, but
the implementation of `compare` still compared signed char, and
claimed \xfe < \x01.

As a consequence, `[^]` on `[\x01-\xfe]` was printed
`[\x80-\xfe] + [\x01-\x7f]`.

+++ vcsn/alphabets/char.hh
@@ -125,7 +125,7 @@ namespace vcsn
     /// Three-way comparison between two letters.
     static int compare(const letter_t l, const letter_t r)
     {
-      return int(l) - int(r);
+      return int(uint8_t(l)) - int(uint8_t(r));
     }
 
     /// Three-way comparison between two words.
